% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_ae_fun.R
\name{hh_evol}
\alias{hh_evol}
\alias{hh_cevo}
\alias{hh_resp}
\alias{hh_auto}
\alias{hh_inte}
\alias{hh_rdif}
\title{Delta method approximation of average evolvability measures}
\usage{
hh_cevo(G)

hh_resp(G)

hh_auto(G)

hh_inte(G)

hh_rdif(G1, G2)
}
\arguments{
\item{G, G1, G2}{Target covariance matrix/matrices.  Assumed validly structured
(symmetric, nonnegative definite), although symmetry is checked.}
}
\description{
These functions calculate Hansen and Houle's (2008, 2009)
delta method approximations for average evolvability measures.

\code{hh_cevo()}: Average conditional evolvability

\code{hh_resp()}: Average respondability

\code{hh_auto()}: Average autonomy

\code{hh_inte()}: Average integration

\code{hh_rdif()}: Average response difference
}
\details{
The notation \eqn{\mathrm{Var}}{Var} in Hansen and Houle's (2008, 2009)
seems to represent average sum of squares from the average (divided by
\eqn{k}, the number of variables) and not the ordinary unbiased estimator
of variance for infinite populations (divided by \eqn{k - 1}).  The
package \code{evolvability} v 2.0.0 incorrectly uses the latter divisor.

Evolvability measures not treated in Hansen and Houle (2008) (flexibility and
response correlation) have not been implemented (although this should be
doable).  All the expressions are for the spherical normal (or uniform on
hypersphyere) distribution of selection gradients.
}
\examples{
## Simple covariance matrices
nv <- 4
G1 <- diag(nv:1)
G2 <- diag(sqrt(1:nv))

######
## Hansen & Houle's (2008) delta method approximations:

## - average conditional evolvability
hh_cevo(G1)

## - average respondability
hh_resp(G1)

## - average autonomy
hh_auto(G1)

## - average response difference
hh_rdif(G1, G2)

}
\references{
Hansen, T. F. and Houle, D. (2008) Measuring and comparing evolvability and
constraint in multivariate characters. \emph{Journal of Evolutionary Biology},
\strong{21}, 1201--1219.
\doi{10.1111/j.1420-9101.2008.01573.x}.

Hansen, T. F. and Houle, D. (2009) Corrigendum [of ``Measuring and comparing
evolvability and constraint in multivariate characters''].
\emph{Journal of Evolutionary Biology},
\strong{22}, 913--915.
\doi{10.1111/j.1420-9101.2009.01715.x}.
}
