% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ae_fun.R
\name{mc_evol}
\alias{mc_evol}
\alias{mc_cevo}
\alias{mc_resp}
\alias{mc_flex}
\alias{mc_auto}
\alias{mc_inte}
\alias{mc_rdif}
\alias{mc_rcor}
\title{Monte Carlo sampling for (average) evolvability measures}
\usage{
mc_evol(nit, G, ...)

mc_cevo(
  nit,
  G,
  mu = rep.int(0, dim(G)[1]),
  Sigma = diag(dim(G)[1]),
  tol_qr = 1e-07,
  ...
)

mc_resp(nit, G, ...)

mc_flex(nit, G, ...)

mc_auto(
  nit,
  G,
  mu = rep.int(0, dim(G)[1]),
  Sigma = diag(dim(G)[1]),
  tol_qr = 1e-07,
  ...
)

mc_inte(nit, G, ...)

mc_rdif(nit, G1, G2, ...)

mc_rcor(nit, G1, G2, ...)
}
\arguments{
\item{nit}{Number of iteration}

\item{G, G1, G2}{Target covariance matrix/matrices.  Assumed validly structured
(symmetric, nonnegative definite), although symmetry is checked.}

\item{...}{Additional arguments are passed to \code{qfratio::rqfr()} or
\code{qfratio::rqfmr()}.  The only useful arguments will be \code{mu} and
\code{Sigma} for the mean vector and covariance matrix, respectively,
of selection gradients.}

\item{mu, Sigma}{Mean vector and covariance matrix, respectively, of selection gradients;
default zero vector and identity matrix.  Passed to
\code{qfratio::rqfr()} or \code{qfratio::rqfmr()}.}

\item{tol_qr}{Tolerance passed to \code{qr()} to determine singularity of \code{G}}
}
\value{
Numeric vector of length \code{nit}
}
\description{
These functions generate a Monte Carlo sample for
evolvability measures, whose average is an estimate of
the average evolvability measure

\code{mc_evol()}: (Average) evolvability

\code{mc_cevo()}: (Average) conditional evolvability

\code{mc_resp()}: (Average) respondability

\code{mc_flex()}: (Average) flexibility

\code{mc_auto()}: (Average) autonomy

\code{mc_inte()}: (Average) integration

\code{mc_rdif()}: (Average) response difference

\code{mc_rcor()}: (Average) response correlation
(aka random skewers correlation)
}
\details{
All these functions are wrappers of either
\code{qfratio::\link[qfratio]{rqfr}()} or
\code{qfratio::\link[qfratio]{rqfmr}()} for random number generation of
simple and multiple ratios of quadratic forms in normal variables.  See
documentations of those functions for technical details.

When \code{G} is singular and the distribution of the selection gradients
is not within its range, the conditional evolvability and autonomy are 0
with probability 1 (see Watanabe, 2022).  Accordingly, \code{mc_cevo()} and
\code{mc_auto()} simply yield a vector of 1's in this case.
}
\examples{
## Simple covariance matrices
nv <- 4
G1 <- diag(nv:1)
G2 <- diag(sqrt(1:nv))
nit <- 1000

## Monte Carlo sample of conditional evolvability
## under spherical distribution of beta,
## and its mean as an estimate of average conditional evolvability
## plus its 95\% CI
mcsample_cevo <- mc_cevo(nit, G1)
mean(mcsample_cevo)
mean(mcsample_cevo) + sd(mcsample_cevo) / sqrt(nit) *
    qt(c(0.025, 0.975), nit - 1)

## Monte Carlo estimate of average response difference
mean(mc_rdif(nit, G1, G2))

## Monte Carlo estimate of average response correlation,
## aka random skewers correlation
mean(mc_rcor(nit, G1, G2))

## Advanced: Monte Carlo estimation of average evolvability under
## non-spherical distribution of selection gradient
mu <- nv:1 / nv
Sigma <- matrix(0.5, nv, nv)
diag(Sigma) <- 1
mean(mc_evol(nit, G1, mu = mu, Sigma = Sigma))

## For singular G, these yield vectors of 1's/0's
## (unless appropriate Sigma is provided)
Gsing <- diag(nv:1 - 1)
mc_cevo(5, Gsing)
mc_auto(5, Gsing)
mc_inte(5, Gsing)

}
\references{
Cheverud, J. M. (1996) Quantitative genetic analysis of cranial morphology
in the cotton-top (\emph{Saguinus oedipus}) and saddle-back(\emph{S. fuscicollis})
tamarines. \emph{Journal of Evolutionary Biology}, \strong{9}, 5--42.
\doi{10.1046/j.1420-9101.1996.9010005.x}.

Cheverud, J. M. and Marroig, G. (2007) Comparing covariance matrices: random
skewers method compared to the common principal components model.
\emph{Genetics and Molecular Biology}, \strong{30}, 461--469.
\doi{10.1590/S1415-47572007000300027}.

Watanabe, J. (2022) Exact expressions and numerical evaluation of average
evolvability measures for characterizing and comparing \strong{G} matrices.
\emph{bioRxiv} preprint, 2022.11.02.514929.
\doi{10.1101/2022.11.02.514929}.
}
\seealso{
\code{\link[qfratio]{rqfr}}
}
